Group 3
We first imported the random function and created 3 variables to give out numbers between 1 and 6. The if function checks whether the 3 random numbers of the variables are equal. If this is the case, the message Yatzy will be shown. We also created a loop that lasts until all 3 variables have the same number.